requirements.txt
psycopg[binary]>=3.2.0
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.0.0
tabulate>=0.9.0